+++ src/projectiles/Proj_rocket.hpp
@@ -7,5 +7,5 @@ class Rocket : public Projectile {
 public:
   Rocket() : Projectile(Faction::None, ProjectileType::ROCKET) {}
 
-  void draw() const override { DrawCircleV(stats.pos, stats.size, RED); }
+  void draw() const override { DrawCircleV(stats.pos, (float)stats.size, RED); }
 };
